Move Group.java into 'group' package
